split helpers for database-backed tests into a separate crate
